Big changes Moved components out of Menu and into App Added CardBody to styledComponents Changed Information and Home|Menu styledComponents to CardBody Added more information describing myself
