--- src/components/App/Menu/Apps.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-class Apps extends React.Component {
-
-  render() {
-    return (
-      <div>
-        Apps
-      </div>
-    );
-  }
-}
-
-export default Apps;
