--- src/components/App/Menu/Home.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-
-import Card from 'sharedComponents/Card';
-import { HomePage } from 'sharedFiles/styledComponents'
-
-const Home = () => {
-
-  return (
-    <HomePage>
-      <Card />
-    </HomePage>
-  );
-}
-
-export default Home;
